+++ src/main/java/com/attornatus/api/controller/openapi/AddressOpenApi.java
@@ -12,12 +12,12 @@ import java.util.List;
 public interface AddressOpenApi {
 
     @ApiOperation("Lista os endereços de uma pessoa através do ID")
-    public List<AddressDTO> listPersonAddress(@ApiParam("ID de uma pessoa") Long personId);
+    public List<AddressDTO> listPersonAddress(@ApiParam("ID de uma pessoa") Integer personId);
 
     @ApiOperation("Cadastra um endereço através do ID da pessoa")
     @ApiResponses({@ApiResponse(code = 404, message = "Pessoa não encontrada", response = Problem.class)})
-    public AddressDTO create(@ApiParam("ID de uma pessoa") @PathVariable Long personId, AddressDTOInput addressDTOInput);
+    public AddressDTO create(@ApiParam("ID de uma pessoa") @PathVariable Integer personId, AddressDTOInput addressDTOInput);
 
     @ApiOperation("Faz escolha do endereço principal através ID da Pessoa e o ID do Endereço")
-    public void addressMain(@ApiParam("ID da pessoa") Long personId, @ApiParam("ID do endereço") Long addressId);
+    public void addressMain(@ApiParam("ID da pessoa") Integer personId, @ApiParam("ID do endereço") Integer addressId);
 }
